Refactor GitHub transport boundaries
